+++ test/flop_phoenix_test.exs
@@ -8,21 +8,27 @@ defmodule Flop.PhoenixTest do
   import Phoenix.LiveViewTest
 
   alias Flop.Meta
+  alias Flop.Phoenix.Pet
   alias Plug.Conn.Query
 
   doctest Flop.Phoenix, import: true
 
   @route_helper_opts [%{}, :pets]
 
-  defp render_pagination(%Meta{} = meta, opts \\ []) do
+  defp render_pagination(%Meta{} = meta, assigns \\ []) do
+    assigns =
+      Keyword.merge(
+        [
+          __changed__: nil,
+          meta: meta,
+          path_helper: &route_helper/3,
+          path_helper_args: @route_helper_opts
+        ],
+        assigns
+      )
+
     (&pagination/1)
-    |> render_component(
-      __changed__: nil,
-      meta: meta,
-      path_helper: &route_helper/3,
-      path_helper_args: @route_helper_opts,
-      opts: opts
-    )
+    |> render_component(assigns)
     |> Floki.parse_fragment!()
   end
 
@@ -36,7 +42,6 @@ defmodule Flop.PhoenixTest do
           meta: %Flop.Meta{flop: %Flop{}},
           path_helper: &route_helper/3,
           path_helper_args: [%{}, :index],
-          opts: [],
           row_func: fn %{name: name}, _opts -> [name] end
         ],
         assigns
@@ -80,7 +85,7 @@ defmodule Flop.PhoenixTest do
       wrapper =
         :meta_on_first_page
         |> build()
-        |> render_pagination(wrapper_attrs: [class: "boo"])
+        |> render_pagination(opts: [wrapper_attrs: [class: "boo"]])
         |> Floki.find("nav")
 
       assert Floki.attribute(wrapper, "aria-label") == ["pagination"]
@@ -92,7 +97,7 @@ defmodule Flop.PhoenixTest do
       wrapper =
         :meta_on_first_page
         |> build()
-        |> render_pagination(wrapper_attrs: [title: "paginate"])
+        |> render_pagination(opts: [wrapper_attrs: [title: "paginate"]])
         |> Floki.find("nav")
 
       assert Floki.attribute(wrapper, "aria-label") == ["pagination"]
@@ -170,8 +175,10 @@ defmodule Flop.PhoenixTest do
         :meta_on_second_page
         |> build()
         |> render_pagination(
-          previous_link_attrs: [class: "prev", title: "p-p-previous"],
-          previous_link_content: tag(:i, class: "fas fa-chevron-left")
+          opts: [
+            previous_link_attrs: [class: "prev", title: "p-p-previous"],
+            previous_link_content: tag(:i, class: "fas fa-chevron-left")
+          ]
         )
 
       assert [link] = Floki.find(html, "a[title='p-p-previous']")
@@ -211,8 +218,10 @@ defmodule Flop.PhoenixTest do
         :meta_on_first_page
         |> build()
         |> render_pagination(
-          previous_link_attrs: [class: "prev", title: "no"],
-          previous_link_content: "Prev"
+          opts: [
+            previous_link_attrs: [class: "prev", title: "no"],
+            previous_link_content: "Prev"
+          ]
         )
         |> Floki.find("span:fl-contains('Prev')")
 
@@ -263,8 +272,10 @@ defmodule Flop.PhoenixTest do
         :meta_on_second_page
         |> build()
         |> render_pagination(
-          next_link_attrs: [class: "next", title: "n-n-next"],
-          next_link_content: tag(:i, class: "fas fa-chevron-right")
+          opts: [
+            next_link_attrs: [class: "next", title: "n-n-next"],
+            next_link_content: tag(:i, class: "fas fa-chevron-right")
+          ]
         )
 
       assert [link] = Floki.find(html, "a[title='n-n-next']")
@@ -306,8 +317,10 @@ defmodule Flop.PhoenixTest do
         :meta_on_last_page
         |> build()
         |> render_pagination(
-          next_link_attrs: [class: "next", title: "no"],
-          next_link_content: "N-n-next"
+          opts: [
+            next_link_attrs: [class: "next", title: "no"],
+            next_link_content: "N-n-next"
+          ]
         )
         |> Floki.find("span:fl-contains('N-n-next')")
 
@@ -373,7 +386,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_second_page
         |> build()
-        |> render_pagination(page_links: :hide)
+        |> render_pagination(opts: [page_links: :hide])
 
       assert Floki.find(html, ".pagination-list") == []
     end
@@ -382,7 +395,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_second_page
         |> build()
-        |> render_pagination(page_links: :hide, event: "paginate")
+        |> render_pagination(opts: [page_links: :hide, event: "paginate"])
 
       assert Floki.find(html, ".pagination-list") == []
     end
@@ -392,7 +405,7 @@ defmodule Flop.PhoenixTest do
         :meta_on_first_page
         |> build()
         |> render_pagination(
-          pagination_list_attrs: [class: "p-list", title: "boop"]
+          opts: [pagination_list_attrs: [class: "p-list", title: "boop"]]
         )
 
       assert [list] = Floki.find(html, "ul.p-list")
@@ -404,7 +417,7 @@ defmodule Flop.PhoenixTest do
         :meta_on_second_page
         |> build()
         |> render_pagination(
-          pagination_link_attrs: [class: "p-link", beep: "boop"]
+          opts: [pagination_link_attrs: [class: "p-link", beep: "boop"]]
         )
 
       assert [link] = Floki.find(html, "a[aria-label='Go to page 1']")
@@ -422,7 +435,7 @@ defmodule Flop.PhoenixTest do
         :meta_on_second_page
         |> build()
         |> render_pagination(
-          current_link_attrs: [class: "link is-active", beep: "boop"]
+          opts: [current_link_attrs: [class: "link is-active", beep: "boop"]]
         )
 
       assert [link] = Floki.find(html, "a[aria-label='Go to page 1']")
@@ -445,7 +458,9 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_second_page
         |> build()
-        |> render_pagination(pagination_link_aria_label: &"On to page #{&1}")
+        |> render_pagination(
+          opts: [pagination_link_aria_label: &"On to page #{&1}"]
+        )
 
       assert [link] = Floki.find(html, "a[aria-label='On to page 1']")
       assert Floki.attribute(link, "class") == ["pagination-link"]
@@ -503,6 +518,53 @@ defmodule Flop.PhoenixTest do
       assert Floki.attribute(next, "data-phx-link") == ["patch"]
       assert Floki.attribute(next, "data-phx-link-state") == ["push"]
       assert Floki.attribute(next, "href") == [expected_url.(3)]
+    end
+
+    test "hides default order and limit" do
+      html =
+        :meta_on_second_page
+        |> build(
+          flop: %Flop{
+            page_size: 20,
+            order_by: [:name],
+            order_directions: [:asc]
+          }
+        )
+        |> render_pagination(for: Pet)
+
+      assert [prev] = Floki.find(html, "a:fl-contains('Previous')")
+      assert [href] = Floki.attribute(prev, "href")
+
+      refute href =~ "page_size="
+      refute href =~ "order_by[]="
+      refute href =~ "order_directions[]="
+    end
+
+    test "does not require path_helper when passing event" do
+      html =
+        (&pagination/1)
+        |> render_component(
+          __changed__: nil,
+          event: "paginate",
+          meta: build(:meta_on_second_page)
+        )
+        |> Floki.parse_fragment!()
+
+      link = Floki.find(html, "a:fl-contains('Previous')")
+
+      assert Floki.attribute(link, "class") == ["pagination-previous"]
+      assert Floki.attribute(link, "phx-click") == ["paginate"]
+      assert Floki.attribute(link, "phx-value-page") == ["1"]
+      assert Floki.attribute(link, "href") == ["#"]
+    end
+
+    test "raises if neither path helper nor event are passed" do
+      assert_raise RuntimeError, fn ->
+        render_component(&pagination/1,
+          __changed__: nil,
+          meta: build(:meta_on_second_page)
+        )
+      end
     end
 
     test "adds filter parameters to links" do
@@ -584,7 +646,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 1
       assert html |> Floki.find(".pagination-link") |> length() == 6
@@ -601,7 +663,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 20, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 1
       assert html |> Floki.find(".pagination-link") |> length() == 6
@@ -617,7 +679,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 12, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 6})
+        |> render_pagination(opts: [page_links: {:ellipsis, 6}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 2
       assert html |> Floki.find(".pagination-link") |> length() == 8
@@ -634,7 +696,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 11, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 2
       assert html |> Floki.find(".pagination-link") |> length() == 7
@@ -651,7 +713,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 10, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 2
       assert html |> Floki.find(".pagination-link") |> length() == 7
@@ -668,7 +730,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 11, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 2
       assert html |> Floki.find(".pagination-link") |> length() == 7
@@ -685,7 +747,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 2, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 1
       assert html |> Floki.find(".pagination-link") |> length() == 6
@@ -701,7 +763,7 @@ defmodule Flop.PhoenixTest do
       html =
         :meta_on_first_page
         |> build(current_page: 18, total_pages: 20)
-        |> render_pagination(page_links: {:ellipsis, 5})
+        |> render_pagination(opts: [page_links: {:ellipsis, 5}])
 
       assert html |> Floki.find(".pagination-ellipsis") |> length() == 1
       assert html |> Floki.find(".pagination-link") |> length() == 6
@@ -718,9 +780,11 @@ defmodule Flop.PhoenixTest do
         :meta_on_first_page
         |> build(current_page: 10, total_pages: 20)
         |> render_pagination(
-          page_links: {:ellipsis, 5},
-          ellipsis_attrs: [class: "dotdotdot", title: "dot"],
-          ellipsis_content: "dot dot dot"
+          opts: [
+            page_links: {:ellipsis, 5},
+            ellipsis_attrs: [class: "dotdotdot", title: "dot"],
+            ellipsis_content: "dot dot dot"
+          ]
         )
 
       assert [el, _] = Floki.find(html, "span[class='dotdotdot']")
@@ -826,11 +890,7 @@ defmodule Flop.PhoenixTest do
     end
 
     test "renders links with click handler" do
-      html =
-        render_table(
-          headers: ["Name", {"Age", :age}],
-          opts: [event: "sort"]
-        )
+      html = render_table(event: "sort", headers: ["Name", {"Age", :age}])
 
       assert [th] = Floki.find(html, "th:fl-contains('Name')")
       assert Floki.children(th, include_text: false) == []
@@ -844,8 +904,9 @@ defmodule Flop.PhoenixTest do
     test "adds phx-target to header links" do
       html =
         render_table(
+          event: "sort",
           headers: ["Name", {"Age", :age}],
-          opts: [event: "sort", target: "here"]
+          target: "here"
         )
 
       assert [_] = Floki.find(html, "th:fl-contains('Name')")
@@ -867,16 +928,40 @@ defmodule Flop.PhoenixTest do
       assert [_] = Floki.find(html, "a:fl-contains('Age')")
       assert [_] = Floki.find(html, "a:fl-contains('Species')")
 
-      # with :for option
+      # with :for assign
       html =
         render_table(
-          headers: [{"Name", :name}, {"Age", :age}, {"Species", :species}],
-          opts: [for: Flop.Phoenix.Pet]
+          for: Flop.Phoenix.Pet,
+          headers: [{"Name", :name}, {"Age", :age}, {"Species", :species}]
         )
 
       assert [_] = Floki.find(html, "a:fl-contains('Name')")
       assert [_] = Floki.find(html, "a:fl-contains('Age')")
       assert [] = Floki.find(html, "a:fl-contains('Species')")
+    end
+
+    test "hides default order and limit" do
+      html =
+        render_table(
+          for: Pet,
+          headers: [{"Name", :name}],
+          meta:
+            build(
+              :meta_on_second_page,
+              flop: %Flop{
+                page_size: 20,
+                order_by: [:name],
+                order_directions: [:desc]
+              }
+            )
+        )
+
+      assert [link] = Floki.find(html, "a:fl-contains('Name')")
+      assert [href] = Floki.attribute(link, "href")
+
+      refute href =~ "page_size="
+      refute href =~ "order_by[]="
+      refute href =~ "order_directions[]="
     end
 
     test "renders order direction symbol" do
@@ -951,13 +1036,15 @@ defmodule Flop.PhoenixTest do
       assert Floki.text(span) == "desc"
     end
 
-    test "renders all items" do
+    test "passes additional assigns to row function and renders all items" do
       html =
         render_table(
           items: [%{name: "George", age: 8}, %{name: "Barbara", age: 2}],
-          opts: [appendix: "-chan"],
+          row_opts: [suffix: "-chan"],
           row_func: fn %{age: age, name: name}, opts ->
-            [name <> opts[:appendix], age]
+            assert Keyword.keys(opts) == [:suffix]
+            suffix = Keyword.fetch!(opts, :suffix)
+            [name <> suffix, age]
           end
         )
 
@@ -974,8 +1061,8 @@ defmodule Flop.PhoenixTest do
     test "allows to set no_results_content" do
       assert render_table(
                items: [],
-               no_results_content: ~E"<div>Nothing!</div>"
-             ) == [{"p", [], ["No results."]}]
+               opts: [no_results_content: ~E"<div>Nothing!</div>"]
+             ) == [{"div", [], ["Nothing!"]}]
     end
 
     test "renders table footer" do
@@ -1009,6 +1096,37 @@ defmodule Flop.PhoenixTest do
     test "does not render table footer if option is not set" do
       html = render_table()
       assert [{"table", [], [{"thead", _, _}, {"tbody", _, _}]}] = html
+    end
+
+    test "does not require path_helper when passing event" do
+      html =
+        (&table/1)
+        |> render_component(
+          __changed__: nil,
+          headers: [{"Name", :name}],
+          items: [%{name: "George"}],
+          meta: %Flop.Meta{flop: %Flop{}},
+          row_func: fn %{name: name}, _opts -> [name] end,
+          event: "sort-table"
+        )
+        |> Floki.parse_fragment!()
+
+      assert [link] = Floki.find(html, "a:fl-contains('Name')")
+
+      assert Floki.attribute(link, "phx-click") == ["sort-table"]
+      assert Floki.attribute(link, "href") == ["#"]
+    end
+
+    test "raises if neither path helper nor event are passed" do
+      assert_raise RuntimeError, fn ->
+        render_component(&table/1,
+          __changed__: nil,
+          headers: [{"Name", :name}],
+          items: [%{name: "George"}],
+          meta: %Flop.Meta{flop: %Flop{}},
+          row_func: fn %{name: name}, _opts -> [name] end
+        )
+      end
     end
   end